+++ mrmustard/math/backend_numpy.py
@@ -124,26 +124,13 @@ class BackendNumpy(BackendBase):
             det = np.linalg.det(matrix)
         return det  # noqa: RET504
 
+    def diagonal(
+        self, array: np.ndarray, offset: int | None, axis1: int | None, axis2: int | None
+    ) -> np.ndarray:
+        return np.diagonal(array, offset=offset, axis1=axis1, axis2=axis2)
+
     def diag(self, array: np.ndarray, k: int = 0) -> np.ndarray:
-        if array.ndim in (1, 2):
-            return np.diag(array, k=k)
-        # fallback into more complex algorithm
-        original_sh = array.shape
-
-        ravelled_sh = (np.prod(original_sh[:-1]), original_sh[-1])
-        array = array.ravel().reshape(*ravelled_sh)
-
-        ret = np.asarray([np.diag(line, k) for line in array])
-        inner_shape = (
-            original_sh[-1] + abs(k),
-            original_sh[-1] + abs(k),
-        )
-        return ret.reshape(original_sh[:-1] + inner_shape)
-
-    def diag_part(self, array: np.ndarray, k: int) -> np.ndarray:
-        ret = np.diagonal(array, offset=k, axis1=-2, axis2=-1)
-        ret.flags.writeable = True
-        return ret
+        return np.diag(array, k=k)
 
     def exp(self, array: np.ndarray) -> np.ndarray:
         return np.exp(array)