+++ .github/CHANGELOG.md
@@ -15,6 +15,10 @@
 * Fixed a bug in `OptimizerJax` that would make the optimizer think it reached a stable optimum upon repeat calls to `minimize`.
 [(#634)](https://github.com/XanaduAI/MrMustard/pull/634)
 
+* Fixed a bug with ``State.fock_distribution`` where batch dimensions and mult-mode states were not handled correctly.
+[(#635)](https://github.com/XanaduAI/MrMustard/pull/635)
+
+
 ---
 
 # Release 1.0.0a1